wip: got robot driving towards rviz goals
I'm not convinced I have line 80 of main.py sending the correct pose
to the robot for navigation. This should be tested more.

Modified main.py to send a target pose on /target, and modified Robot.py
to accept targets and immediately navigate to them.

Currently crashes into walls sometimes in testing