got driving working again, A* optimization broken
Currently drives to the next cell along the optimal path to the goal given
through rviz. If optimize_paths in A_Star.py is fixed, this will be ready
for the next step (getting main to send the closest frontier cell in place of
the rviz provided goal)

We think the A* start and goal cells are not actually being lined up to the
grid, so the first movement along the path is usually adjusting into the center
of the nearest cell. We might need to change line 80 in main.py to grab the
third cell along the path, rather than the second if that issue persists.